:recycle: moving and dividing employee to employee and employee-list
